// src/models/_fndds-models/db.d.ts
export interface nutDataAttribute {
  ndbNo: string;
  nutrNo: string;
  nutrVal: number;
  numDataPts?: number;
  stdError?: number;
  srcCd?: string;
  derivCd?: string;
  refNdbNo?: string;
  addNutrMark?: string;
  numStudies?: number;
  min?: number;
  max?: number;
  df?: number;
  lowEb?: number;
  upEb?: number;
  statCmt?: string;
  addModDate?: string;
}

export interface nutDataInstance extends Sequelize.Instance<nutDataAttribute>, nutDataAttribute {
}